refactor(net): move packages into net dir
rename and move libp2p package to net/host, move client package to net/client
