+++ pkg/internal/itest/client_query_test.go
@@ -8,8 +8,8 @@ import (
 	"github.com/filecoin-project/go-address"
 	retrievaltypes "github.com/filecoin-project/go-retrieval-types"
 	"github.com/filecoin-project/go-state-types/abi"
-	"github.com/filecoin-project/lassie/pkg/client"
 	"github.com/filecoin-project/lassie/pkg/internal/itest/mocknet"
+	"github.com/filecoin-project/lassie/pkg/net/client"
 	"github.com/ipfs/go-cid"
 	datastore "github.com/ipfs/go-datastore"
 	dss "github.com/ipfs/go-datastore/sync"
